Adopt "tables inherit color from body quirk" from Gecko (#333)
Already landed upstream:
https://bugzilla.mozilla.org/show_bug.cgi?id=2023380

Servo PR: https://github.com/servo/servo/pull/43368
