+++ style/color/mod.rs
@@ -638,6 +638,30 @@ impl AbsoluteColor {
             self.alpha(),
         )
     }
+
+    /// Convert a color value to `nscolor`.
+    pub fn to_nscolor(&self) -> u32 {
+        let srgb = self.to_color_space(ColorSpace::Srgb);
+        u32::from_le_bytes([
+            (srgb.components.0 * 255.0).round() as u8,
+            (srgb.components.1 * 255.0).round() as u8,
+            (srgb.components.2 * 255.0).round() as u8,
+            (srgb.alpha * 255.0).round() as u8,
+        ])
+    }
+
+    /// Convert a given `nscolor` to a Servo AbsoluteColor value.
+    pub fn from_nscolor(color: u32) -> Self {
+        let [r, g, b, a] = color.to_le_bytes();
+        Self::srgb_legacy(r, g, b, a as f32 / 255.0)
+    }
+}
+
+#[test]
+fn from_nscolor_should_be_in_legacy_syntax() {
+    let result = AbsoluteColor::from_nscolor(0x336699CC);
+    assert!(result.flags.contains(ColorFlags::IS_LEGACY_SRGB));
+    assert!(result.is_legacy_syntax());
 }
 
 impl From<PredefinedColorSpace> for ColorSpace {
