+++ style/device/mod.rs
@@ -4,6 +4,7 @@
 
 //! Media-query device and expression representation.
 
+use crate::color::AbsoluteColor;
 use crate::custom_properties::CssEnvironment;
 #[cfg(feature = "servo")]
 use crate::derives::*;
@@ -88,6 +89,12 @@ pub struct Device {
     /// The CssEnvironment object responsible of getting CSS environment
     /// variables.
     environment: CssEnvironment,
+    /// The body text color, stored as an `nscolor`, used for the "tables
+    /// inherit from body" quirk.
+    ///
+    /// <https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk>
+    #[cfg_attr(feature = "servo", ignore_malloc_size_of = "Pure stack type")]
+    body_text_color: AtomicU32,
 
     /// Extra Gecko-specific or Servo-specific data.
     extra: ExtraDeviceData,
@@ -279,5 +286,18 @@ impl Device {
     /// Returns whether font metrics have been queried.
     pub fn used_font_metrics(&self) -> bool {
         self.used_font_metrics.load(Ordering::Relaxed)
+    }
+
+    /// Returns the body text color.
+    pub fn body_text_color(&self) -> AbsoluteColor {
+        AbsoluteColor::from_nscolor(self.body_text_color.load(Ordering::Relaxed))
+    }
+
+    /// Sets the body text color for the "inherit color from body" quirk.
+    ///
+    /// <https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk>
+    pub fn set_body_text_color(&self, color: AbsoluteColor) {
+        self.body_text_color
+            .store(color.to_nscolor(), Ordering::Relaxed)
     }
 }
