+++ style/gecko/mod.rs
@@ -11,7 +11,6 @@ pub mod arc_types;
 pub mod conversions;
 pub mod data;
 pub mod media_features;
-pub mod media_queries;
 pub mod pseudo_element;
 pub mod restyle_damage;
 pub mod selector_parser;
@@ -19,5 +18,4 @@ pub mod snapshot;
 pub mod snapshot_helpers;
 pub mod traversal;
 pub mod url;
-pub mod values;
 pub mod wrapper;
